Adding workflow parameters panel
