+++ libs/designer/src/lib/core/index.ts
@@ -5,3 +5,4 @@ export type { RootState, AppDispatch } from './store';
 export { store } from './store';
 export { discardAllChanges } from './state/workflow/workflowSlice';
 export { serializeWorkflow } from './actions/bjsworkflow/serializer';
+export { switchToWorkflowParameters } from './state/panel/panelSlice';
