+++ libs/designer/src/lib/core/state/panel/panelSlice.ts
@@ -49,6 +49,7 @@ export const panelSlice = createSlice({
     ) => {
       state.collapsed = false;
       state.isDiscovery = true;
+      state.isWorkflowParameters = false;
       state.relationshipIds = action.payload.relationshipIds;
       state.selectedNode = action.payload.nodeId;
       state.isParallelBranch = action.payload?.isParallelBranch ?? false;
@@ -56,6 +57,8 @@ export const panelSlice = createSlice({
     },
     selectOperationGroupId: (state, action: PayloadAction<string>) => {
       state.selectedOperationGroupId = action.payload;
+      state.isDiscovery = false;
+      state.isWorkflowParameters = false;
     },
     switchToOperationPanel: (state, action: PayloadAction<string>) => {
       state.selectedNode = action.payload;
@@ -141,6 +144,7 @@ export const {
   isolateTab,
   selectPanelTab,
   setTabError,
+  switchToWorkflowParameters,
 } = panelSlice.actions;
 
 export default panelSlice.reducer;
