+++ libs/designer/src/lib/core/utils/validation.ts
@@ -14,8 +14,8 @@ const regex = {
   datetime:
     /^(\d{4})-(0[1-9]|1[012])-(0[1-9]|[12]\d|3[01])(?:[\sT])([01]\d|2[0-3]):([0-5]\d)(?::([0-5]\d|60))?(\.\d+)?(([Zz])|([+|-]([01]\d|2[0-3])))?$/,
   double: /^(?:[-+])?([0-9]*(\.[0-9]+([eE](?:[-+])[0-9]+)?)?)$/,
-  // eslint-disable-next-line no-empty-character-class
   email:
+    // eslint-disable-next-line no-empty-character-class
     /^([!#-'*+/-9=?A-Z^-~-]+(\.[!#-'*+/-9=?A-Z^-~-]+)*|"([]!#-[^-~ \t]|(\\[\t -~]))+")@([!#-'*+/-9=?A-Z^-~-]+(\.[!#-'*+/-9=?A-Z^-~-]+)*|\[[\t -Z^-~]*])$/,
   integer: /^(?:[-+])?([0-9]+)$/,
   phone: /^(\+)?(?:[0-9]{5,15})$/,
